+++ frontend/src/repositories/HeroRepository.js
@@ -1,0 +1,8 @@
+// frontend/src/repositories/HeroRepository.js
+import { heroMock } from "../mocks/hero.mock.js";
+
+export const HeroRepository = {
+    async getHero() {
+        return heroMock;
+    }
+}